devop: 🔧 move marketData functions to helper file
